Added java doc, moved the PaymentService.kt to the services folder, removed unused imports and useless semicolon in the SaveButton.kt and fixed a small typo
