+++ app/src/main/java/com/rentminder/SaveButton.kt
@@ -1,36 +1,25 @@
 package com.rentminder
 
 import android.widget.Toast
-import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.core.LinearOutSlowInEasing
-import androidx.compose.animation.core.animate
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.border
-import androidx.compose.foundation.indication
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.*
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 
+
+/**
+ * This is the button for saving any edits or initial inputs
+ */
 @Composable
 fun SaveRemindButton(enabled: Boolean, inputEdited: MutableState<Boolean>) {
     val context = LocalContext.current
@@ -82,11 +71,14 @@ fun SaveRemindButton(enabled: Boolean, inputEdited: MutableState<Boolean>) {
     }
 }
 
+/**
+ * This is a preview of the save remind button
+ */
 @ExperimentalMaterialApi
 @Composable
 @Preview
 private fun SaveRemindButtonPreview() {
-    val enabled = false;
+    val enabled = false
     val inputEdited = remember { mutableStateOf(false) }
     SaveRemindButton(enabled, inputEdited)
 }
