+++ app/src/main/java/com/rentminder/service/PaymentService.kt
@@ -1,4 +1,5 @@
-package com.rentminder
+package com.rentminder.service
+import com.rentminder.RetrofitClientInstance
 import com.rentminder.dao.PaymentDao
 import com.rentminder.dto.Payment
 import kotlinx.coroutines.Dispatchers
@@ -6,6 +7,10 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.withContext
 import retrofit2.awaitResponse
 class PaymentService {
+
+    /**
+     * Function for fetching the payment and connecting it to our Dao
+     */
     suspend fun fetchPayment(): List<Payment>? {
         return withContext(Dispatchers.IO) {
             val service = RetrofitClientInstance.retrofitInstance?.create(PaymentDao::class.java)
